requirements.txt
Django==5.0.6
djangorestframework==3.15.2
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
mysqlclient==2.2.4
gunicorn==22.0.0
setuptools
whitenoise==6.5.0
django-filter>=24.2
